Finish merge and get tests passing
